test: move estimator builder test to density_estimator_test file
